removed old security config file, and moved firebase config to utils
