+++ src/main/java/com/engagement/util/FirebaseConfig.java
@@ -1,4 +1,4 @@
-package com.engagement;
+package com.engagement.util;
 
 import java.io.IOException;
