--- src/main/java/com/engagement/util/SecurityConfigurer.java
@@ -1,33 +1,0 @@
-package com.engagement.util;
-
-//@Configuration
-//@EnableWebSecurity
-//public class SecurityConfigurer extends WebSecurityConfigurerAdapter {
-//    private static final String[] AUTH_WHITELIST = {
-//    		// --swagger ui
-//    		"/swagger-resources/**",
-//    		"/swagger-ui.html",
-//    		"/v2/api-docs",
-//    		"/webjars/**",
-//    		"/swagger-ui/**"
-//    };
-//    
-//	/**
-//	 * Overriding the default implementation of the filter given to us 
-//	 * by Spring Security.  We disable CORS restrictions, and make it to where
-//	 * any request that comes into the server is authenticated first before 
-//	 * reaching our resources.  The only requests that are not authenticated
-//	 * pertain to Swagger-ui and its resouces.
-//	 */
-//	@Override
-//    protected void configure(HttpSecurity http) throws Exception {
-//    	http.authorizeRequests().requestMatchers(CorsUtils::isPreFlightRequest).permitAll()
-//    	.antMatchers(AUTH_WHITELIST).permitAll();
-//    	http.cors().and().csrf().disable().authorizeRequests(authorize -> authorize
-//                .anyRequest().authenticated()
-//                )
-//                .oauth2ResourceServer(OAuth2ResourceServerConfigurer::jwt);
-//        
-//    }
-    
-//}
